Inception family of models

This is the API reference for the Inception family of models supported by Metalhead.jl.

The higher-level model constructors

Metalhead.GoogLeNet — Type

GoogLeNet(; pretrain::Bool = false, inchannels::Integer = 3, nclasses::Integer = 1000)

Create an Inception-v1 model (commonly referred to as GoogLeNet) (reference).

Arguments

pretrain: set to true to load the model with pre-trained weights for ImageNet

nclasses: the number of output classes

batchnorm: set to true to use batch normalization after each convolution

bias: set to true to use bias in the convolution layers

Warning

GoogLeNet does not currently support pretrained weights.

See also Metalhead.googlenet.

Metalhead.Inceptionv3 — Type

Inceptionv3(; pretrain::Bool = false, inchannels::Integer = 3, nclasses::Integer = 1000)

Create an Inception-v3 model (reference).

Arguments

pretrain: set to true to load the pre-trained weights for ImageNet

inchannels: number of input channels

nclasses: the number of output classes

Warning

Inceptionv3 does not currently support pretrained weights.

See also Metalhead.inceptionv3.

Metalhead.Inceptionv4 — Type

Inceptionv4(; pretrain::Bool = false, inchannels::Integer = 3,
            nclasses::Integer = 1000)

Creates an Inceptionv4 model. (reference)

Arguments

pretrain: set to true to load the pre-trained weights for ImageNet

inchannels: number of input channels.

nclasses: the number of output classes.

Warning

Inceptionv4 does not currently support pretrained weights.

See also Metalhead.inceptionv4.

Metalhead.InceptionResNetv2 — Type

InceptionResNetv2(; pretrain::Bool = false, inchannels::Integer = 3, 
                  nclasses::Integer = 1000)

Creates an InceptionResNetv2 model. (reference)

Arguments

pretrain: set to true to load the pre-trained weights for ImageNet

inchannels: number of input channels.

nclasses: the number of output classes.

Warning

InceptionResNetv2 does not currently support pretrained weights.

See also Metalhead.inceptionresnetv2.

Metalhead.Xception — Type

Xception(; pretrain::Bool = false, inchannels::Integer = 3, nclasses::Integer = 1000)

Creates an Xception model. (reference)

Arguments

pretrain: set to true to load the pre-trained weights for ImageNet.

inchannels: number of input channels.

nclasses: the number of output classes.

Warning

Xception does not currently support pretrained weights.

See also Metalhead.xception.

The mid-level functions

Metalhead.googlenet — Function

googlenet(; dropout_prob = 0.4, inchannels::Integer = 3, nclasses::Integer = 1000)

Create an Inception-v1 model (commonly referred to as GoogLeNet) (reference).

Arguments

dropout_prob: the dropout probability in the classifier head. Set to nothing to disable dropout.

inchannels: the number of input channels

nclasses: the number of output classes

batchnorm: set to true to include batch normalization after each convolution

bias: set to true to use bias in the convolution layers

Metalhead.inceptionv3 — Function

inceptionv3(; dropout_prob = 0.2, inchannels::Integer = 3, nclasses::Integer = 1000)

Create an Inception-v3 model (reference).

Arguments

dropout_prob: the dropout probability in the classifier head. Set to nothing to disable dropout.

inchannels: number of input feature maps

nclasses: the number of output classes

Metalhead.inceptionv4 — Function

inceptionv4(; dropout_prob = nothing, inchannels::Integer = 3, nclasses::Integer = 1000)

Create an Inceptionv4 model. (reference)

Arguments

dropout_prob: probability of dropout in classifier head. Set to nothing to disable dropout.

inchannels: number of input channels.

nclasses: the number of output classes.

Metalhead.inceptionresnetv2 — Function

inceptionresnetv2(; inchannels::Integer = 3, dropout_prob = nothing, nclasses::Integer = 1000)

Creates an InceptionResNetv2 model. (reference)

Arguments

dropout_prob: probability of dropout in classifier head. Set to nothing to disable dropout.

inchannels: number of input channels.

nclasses: the number of output classes.

Metalhead.xception — Function

xception(; dropout_prob = nothing, inchannels::Integer = 3, nclasses::Integer = 1000)

Creates an Xception model. (reference)

Arguments

dropout_prob: probability of dropout in classifier head. Set to nothing to disable dropout.

inchannels: number of input channels.

nclasses: the number of output classes.