Cora

MLDatasets.Cora — Module

Cora

The Cora citation network dataset from Ref. [1]. Nodes represent documents and edges represent citation links. Each node has a predefined feature with 1433 dimensions. The dataset is designed for the node classification task. The task is to predict the category of certain paper. The dataset is retrieved from Ref. [2].

Statistics

Nodes: 2708

Edges: 10556

Number of Classes: 7

Label split:

Train: 140

Val: 500

Test: 1000

The split is the one used in the original paper [1] and doesn't consider all nodes.

Interface

Cora.dataset

References

[1]: Deep Gaussian Embedding of Graphs: Unsupervised Inductive Learning via Ranking [2]: [Planetoid](https://github.com/kimiyoung/planetoid

API reference

MLDatasets.Cora.dataset — Function

dataset(; dir=nothing, reverse_edges=true)

Retrieve the Cora dataset. The output is a named tuple with fields

julia> keys(Cora.dataset())
(:node_features, :node_labels, :adjacency_list, :train_indices, :val_indices, :test_indices, :num_classes, :num_nodes, :num_edges, :directed)

In particular, adjacency_list is a vector of vector, where adjacency_list[i] will contain the neighbors of node i through outgoing edges.

If reverse_edges=true, the graph will contain the reverse of each edge and the graph will be undirected.

See also Cora.

Usage Examples

using MLDatasets: Cora

data = Cora.dataset()
train_labels = data.node_labels[data.train_indices]