Mutagenesis

MLDatasets.Mutagenesis — Module

Mutagenesis

Website: https://relational.fit.cvut.cz/dataset/Mutagenesis License: CC0

The Mutagenesis dataset comprises 188 molecules trialed for mutagenicity on Salmonella typhimurium, available from relational.fit.cvut.cz and CTUAvastLab/datasets.

Train, test and validation data can be loaded using following code. The withenv("DATADEPS_ALWAYS_ACCEPT"=>"true") disables the accept download prompt.

julia> using MLDatasets: Mutagenesis

julia> train_x, train_y = withenv("DATADEPS_ALWAYS_ACCEPT"=>"true") do; Mutagenesis.traindata(); end;

julia> test_x, test_y = Mutagenesis.testdata();

julia> val_x, val_y = Mutagenesis.valdata();

julia> train_x[1]
JSON3.Object{Base.CodeUnits{UInt8, String}, SubArray{UInt64, 1, Vector{UInt64}, Tuple{UnitRange{Int64}}, true}} with 6 entries:
  :ind1      => 1
  :inda      => 0
  :logp      => 4.23
  :lumo      => -1.246
  :mutagenic => 1
  :atoms     => JSON3.Object[{…

julia> train_y[1]
1