Dataset Containers

MLDatasets.jl contains several reusable data containers for accessing datasets in common storage formats. This feature is a work-in-progress and subject to change.

MLDatasets.FileDataset — Type

FileDataset([loadfn = FileIO.load,] paths)
FileDataset([loadfn = FileIO.load,] dir, pattern = "*", depth = 4)

Wrap a set of file paths as a dataset (traversed in the same order as paths). Alternatively, specify a dir and collect all paths that match a glob pattern (recursively globbing by depth). The glob order determines the traversal order.

MLDatasets.TableDataset — Type

TableDataset(table)
TableDataset(path::AbstractString)

Wrap a Tables.jl-compatible table as a dataset container. Alternatively, specify the path to a CSV file directly to load it with CSV.jl + DataFrames.jl.

MLDatasets.HDF5Dataset — Type

HDF5Dataset(file::AbstractString, paths)
HDF5Dataset(fid::HDF5.File, paths::Union{HDF5.Dataset, Vector{HDF5.Dataset}})
HDF5Dataset(fid::HDF5.File, paths::Union{AbstractString, Vector{<:AbstractString}})
HDF5Dataset(fid::HDF5.File, paths::Union{HDF5.Dataset, Vector{HDF5.Dataset}}, shapes)

Wrap several HDF5 datasets (paths) as a single dataset container. Each dataset p in paths should be accessible as fid[p]. Calling getobs on a HDF5Dataset returns a tuple with each element corresponding to the observation from each dataset in paths. See close(::HDF5Dataset) for closing the underlying HDF5 file pointer.

For array datasets, the last dimension is assumed to be the observation dimension. For scalar datasets, the stored value is returned by getobs for any index.

Base.close — Method

close(dataset::HDF5Dataset)

Close the underlying HDF5 file pointer for dataset.

MLDatasets.JLD2Dataset — Type

JLD2Dataset(file::AbstractString, paths)
JLD2Dataset(fid::JLD2.JLDFile, paths::Union{String, Vector{String}})

Wrap several JLD2 datasets (paths) as a single dataset container. Each dataset p in paths should be accessible as fid[p]. Calling getobs on a JLD2Dataset is equivalent to mapping getobs on each dataset in paths. See close(::JLD2Dataset) for closing the underlying JLD2 file pointer.

Base.close — Method

close(dataset::JLD2Dataset)

Close the underlying JLD2 file pointer for dataset.

MLDatasets.CachedDataset — Type

CachedDataset(source, cachesize = numbobs(source))
CachedDataset(source, cacheidx = 1:numbobs(source))
CachedDataset(source, cacheidx, cache)

Wrap a source data container and cache cachesize samples in memory. This can be useful for improving read speeds when source is a lazy data container, but your system memory is large enough to store a sizeable chunk of it.

By default the observation indices 1:cachesize are cached. You can manually pass in a set of cacheidx as well.

See also make_cache for customizing the default cache creation for source.

MLDatasets.make_cache — Function

make_cache(source, cacheidx)

Return a in-memory copy of source at observation indices cacheidx. Defaults to getobs(source, cacheidx).